faust_app/requirements.txt
faust-streaming==0.11.3
sortedcontainers==2.4.0
rocksdict==0.3.29
pydantic-settings==2.13.1